Split v2 serializer and deserializer into their own files
